requirements.txt
requests>=2.25.1
boto3>=1.26.0
python-dotenv>=1.0.1
PyYAML>=6.0.0
